squash: fix circular dep
